Fixes bugs

+++ veamy/include/feamy/config/FeamyConfig.h
@@ -9,7 +9,7 @@ private:
 public:
     void setNumberOfGaussPoints(int gauss);
 
-    double getNumberOfGaussPoints();
+    int getNumberOfGaussPoints();
 
     static  FeamyConfig* instance();
 };
